Rename "bytesrw" to "mempipe".
