+++ pylib-v1/ailets/cons/mempipe.py
@@ -5,7 +5,7 @@ from typing import Optional
 from .atyping import IAsyncReader, IAsyncWriter
 from .notification_queue import INotificationQueue, NotificationQueue
 
-logger = logging.getLogger("ailets.io")
+logger = logging.getLogger("ailets.mempipe")
 
 
 class Writer(IAsyncWriter):
@@ -22,11 +22,11 @@ class Writer(IAsyncWriter):
         self.queue = queue
         self.debug_hint = debug_hint
         self.closed = False
-        self.queue.whitelist(handle, f"BytesWR.Writer {debug_hint}")
+        self.queue.whitelist(handle, f"MemPipe.Writer {debug_hint}")
 
     def __str__(self) -> str:
         return (
-            f"BytesWR.Writer(handle={self.handle}, "
+            f"MemPipe.Writer(handle={self.handle}, "
             f"closed={self.closed}, "
             f"tell={self.tell()}, "
             f"hint={self.debug_hint})"
@@ -85,7 +85,7 @@ class Reader(IAsyncReader):
             data = self.writer.buffer[slice(self.pos, end_pos)]
             if logger.isEnabledFor(logging.DEBUG):
                 logger.debug(
-                    "BytesWR.Reader.read: handle=%s, old pos=%s, new pos=%s",
+                    "MemPipe.Reader.read: handle=%s, old pos=%s, new pos=%s",
                     self.handle,
                     self.pos,
                     end_pos,
@@ -101,14 +101,14 @@ class Reader(IAsyncReader):
         with lock:
             if self._should_wait_with_autoclose():
                 await self.writer.queue.wait_unsafe(
-                    self.writer.handle, f"BytesWR.Reader {self.handle}"
+                    self.writer.handle, f"MemPipe.Reader {self.handle}"
                 )
                 lock.acquire()
         if self.writer.closed:
             self.close()
 
 
-class BytesWR:
+class MemPipe:
     def __init__(
         self,
         writer_handle: int,
@@ -123,12 +123,12 @@ class BytesWR:
 
     def get_reader(self, handle: int) -> IAsyncReader:
         logger.debug(
-            "BytesWR.get_reader: %s for the writer %s", handle, self.writer.handle
+            "MemPipe.get_reader: %s for the writer %s", handle, self.writer.handle
         )
         return Reader(handle, self.writer)
 
     def __str__(self) -> str:
-        return f"BytesWR(writer={self.writer})"
+        return f"MemPipe(writer={self.writer})"
 
 
 def main() -> None:
@@ -154,7 +154,7 @@ def main() -> None:
 
     async def main() -> None:
         queue = NotificationQueue()
-        wr = BytesWR(0, queue, "main")
+        wr = MemPipe(0, queue, "main")
         lib_writer = wr.get_writer()
         lib_reader1 = wr.get_reader(1)
         lib_reader2 = wr.get_reader(2)
